+++ deltafi-core/src/main/java/org/deltafi/core/metrics/statsd/StatsdClient.java
@@ -15,7 +15,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.deltafi.common.metrics.statsd;
+package org.deltafi.core.metrics.statsd;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;